Relocate MetadataService
