+++ packages/server/src/attestation/verifications/verifyPacked.ts
@@ -15,7 +15,7 @@ import convertX509CertToPEM from '../../helpers/convertX509CertToPEM';
 import getCertificateInfo from '../../helpers/getCertificateInfo';
 import verifySignature from '../../helpers/verifySignature';
 import decodeCredentialPublicKey from '../../helpers/decodeCredentialPublicKey';
-import MetadataService from '../../metadata/metadataService';
+import MetadataService from '../../services/metadataService';
 import verifyAttestationWithMetadata from '../../metadata/verifyAttestationWithMetadata';
 
 type Options = {
