+++ packages/server/src/index.ts
@@ -6,7 +6,7 @@ import generateAttestationOptions from './attestation/generateAttestationOptions
 import verifyAttestationResponse from './attestation/verifyAttestationResponse';
 import generateAssertionOptions from './assertion/generateAssertionOptions';
 import verifyAssertionResponse from './assertion/verifyAssertionResponse';
-import MetadataService from './metadata/metadataService';
+import MetadataService from './services/metadataService';
 
 export {
   generateAttestationOptions,
